+++ src/main/java/cn/edu/guet/service/IWaterBuyService.java
@@ -13,6 +13,6 @@ public interface IWaterBuyService {
     public int countWater();
     public List<Student> getStudent();
     public List<Room> getAllRoom(String buid);
-    public List<Waterorder> insertOrder(String woId, String buid, int woNum, String woRoom, int woMoney, String woUser);
+    public List<Waterorder> insertOrder(String woId, String buid, int woNum, String woRoom, double woMoney, String woUser);
     public List<Waterorder> getAllOrder(String woUser);
 }
